fix values and scroll

+++ lib/models/simulator_configuration_model.dart
@@ -14,7 +14,7 @@ class Product {
   int prazoMaximo;
   int carenciaMinima;
   int carenciaMaxima;
-  int? bonusDia;
+  double? bonusDia;
   String nome;
   String descricao;
   DateTime createdAt;
